Allow compilation from any base directory
The precompiled lcc.exe had \stvoy\Code-DM\lcc\bin\cpp and
\stvoy\Code-DM\lcc\bin\rcc as hard coded strings in it, which basically
assumed that you would check out this repository with name "stvoy" in
your root directory. Modifying these strings in the binary to the
relative paths ..\..\lcc\..\.\lcc\bin\cpp and ..\..\lcc\..\.\lcc\bin\rcc
fixes that and allows to put the repository anywhere, as long as the
working directory while running lcc.exe is two folders deep relative to
Code-DM, which is the case for the .bat files which run lcc.exe from
Code-DM/cgame/vm, Code-DM/game/vm and Code-DM/ui/vm. The superfluous
steps in the relative paths are to keep the length of the strings the
same, otherwise the binary would break.

Similarly, the q3asm file would use \stvoy\baseef\vm\... as output
directory for the .qvm files, which we change to just outputting in the
current working directory that is Code-DM/.../vm for the .bat files.